Allow addition of dataset without resources
Temporary fix for https://github.com/ckan/ckanext-scheming/issues/377:

When ckan is configured with
`CKAN__DATASET__CREATE_ON_UI_REQUIRES_RESOURCES=False`, you should be
able to create a dataset without any resources but the scheming plugin
overrides this as it sets the state to draft until any resources are
added. This temporary fix means that the resources step is omitted from
the add dataset form and a catalog entry is created with no resources.

Not a solution worth actually using - the correct fix would allow the
option for something like "save without adding resources" on the
resource page instead of skipping over the resources. I haven't done
that since I don't need it for what I'm doing right now and it sounds
like the maintainers are on fixing the issue.